Moved People into its own file
